+++ src/index.ts
@@ -26,6 +26,7 @@ export type {
   VehicleDefinitions,
   WordDefinitions,
 } from './definitions';
+export { FakerError } from './errors/faker-error';
 export type { FakerOptions, UsableLocale, UsedLocales } from './faker';
 export { Gender } from './name';
 export type { GenderType } from './name';