+++ src/locales/ar/address/index.ts
@@ -9,7 +9,6 @@ import city_name from './city_name';
 import country from './country';
 import default_country from './default_country';
 import postcode from './postcode';
-import postcode_by_state from './postcode_by_state';
 import secondary_address from './secondary_address';
 import state from './state';
 import street_address from './street_address';
@@ -23,7 +22,6 @@ const address = {
   country,
   default_country,
   postcode,
-  postcode_by_state,
   secondary_address,
   state,
   street_address,