+++ src/locales/fa/address/index.ts
@@ -9,7 +9,6 @@ import city_name from './city_name';
 import city_prefix from './city_prefix';
 import city_suffix from './city_suffix';
 import country from './country';
-import country_code from './country_code';
 import default_country from './default_country';
 import direction from './direction';
 import postcode from './postcode';
@@ -27,7 +26,6 @@ const address = {
   city_prefix,
   city_suffix,
   country,
-  country_code,
   default_country,
   direction,
   postcode,