+++ src/locales/he/address/index.ts
@@ -8,14 +8,11 @@ import city from './city';
 import city_prefix from './city_prefix';
 import city_suffix from './city_suffix';
 import country from './country';
-import country_code from './country_code';
-import country_code_alpha_3 from './country_code_alpha_3';
 import county from './county';
 import default_country from './default_country';
 import direction from './direction';
 import direction_abbr from './direction_abbr';
 import postcode from './postcode';
-import postcode_by_state from './postcode_by_state';
 import secondary_address from './secondary_address';
 import state from './state';
 import state_abbr from './state_abbr';
@@ -30,14 +27,11 @@ const address = {
   city_prefix,
   city_suffix,
   country,
-  country_code,
-  country_code_alpha_3,
   county,
   default_country,
   direction,
   direction_abbr,
   postcode,
-  postcode_by_state,
   secondary_address,
   state,
   state_abbr,